Split composer staging boundaries
