+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.20",
-  "tag": "v3.4.20",
+  "version": "3.4.21",
+  "tag": "v3.4.21",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.19 <3.4.20"
+      ">=3.4.20 <3.4.21"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.19 first."
+    "message": "Update to v3.4.20 first."
   }
 }
